requirements.txt
pytest==7.2.0
selenium==4.7.2
allure-pytest==2.12.0
requests==2.28.1
pytest-xdist==3.1.0